Permit duplicate vector register element updates
Because Handcar relies on register access methods to determine when to send vector register element
updates, it is very difficult to carefully control them. This results in some spurious, but harmless
updates being sent. The updates are consolidated when translating them to physical register
boundaries, so the end result is as expected.

Signed-off-by: Noah Sherrill <nsherril@futurewei.com>